refactor: Break up V2 Login View
